fix: rename Sponsers to Sponsors
